requirements.txt
awswrangler==3.6.0
boto3==1.34.0
botocore==1.34.162
certifi==2026.2.25
charset-normalizer==3.4.6
Faker==22.0.0
greenlet==3.3.2
idna==3.11
jmespath==1.1.0
numpy==1.26.4
packaging==23.2
pandas==2.1.4
psycopg2-binary==2.9.9
pyarrow==23.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2026.1.post1
requests==2.31.0
s3transfer==0.9.0
six==1.17.0
SQLAlchemy==2.0.25
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
